% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correlatedSobol2Sequence}
\alias{correlatedSobol2Sequence}
\title{Generate correlated 2D Sobol' quasirandom sequence}
\usage{
correlatedSobol2Sequence(rho, n, skip = 0L)
}
\arguments{
\item{rho}{correlation}

\item{n}{number of variates to sample}

\item{skip}{number of variates to skip (actual number skipped will be largest power of 2 less than k)}
}
\value{
a n-by-2 matrix of uniform correlated probabilities in (0,1).
}
\description{
Generate correlated 2D Sobol' quasirandom sequence
}
\examples{
correlatedSobol2Sequence(0.2, 1000)
}
