% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{flatten}
\alias{flatten}
\title{Convert multidimensional array of counts per state into table form. Each row in the table corresponds to one individual}
\usage{
flatten(stateOccupancies)
}
\arguments{
\item{stateOccupancies}{an arbitrary-dimension array of (integer) state occupation counts.}
}
\value{
a DataFrame with columns corresponding to category values and rows corresponding to individuals.
}
\description{
This function
}
