% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanleague.R
\docType{package}
\name{humanleague}
\alias{humanleague}
\alias{humanleague-package}
\title{humanleague}
\description{
R package for synthesising populations from aggregate and (optionally) seed data
}
\details{
See README.md for detailed information and examples.
}
\section{Overview}{

}

\section{Functions}{

\code{\link{flatten}}

\code{\link{ipf}}

\code{\link{prob2IntFreq}}

\code{\link{qis}}

\code{\link{qisi}}

\code{\link{sobolSequence}}

\code{\link{synthPop}}

\code{\link{synthPopG}}

\code{\link{unitTest}}
}

