% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{integerise}
\alias{integerise}
\title{Generate integer population from a fractional one where the 1-d partial sums along each axis have an integral total}
\usage{
integerise(population)
}
\arguments{
\item{population}{a numeric vector of state occupation probabilities. Must sum to unity (to within double precision epsilon)}
}
\value{
an integer vector of frequencies that sums to pop.
}
\description{
This function will generate the closest integer array to the fractional population provided, preserving the sums in every dimension.
}
\examples{
prob2IntFreq(c(0.1,0.2,0.3,0.4), 11)
}
