% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pop}
\alias{pop}
\title{Generate a 2D population matrix}
\usage{
pop(marginal0, marginal1, maxAttempts = 4L)
}
\arguments{
\item{marginal0}{an integer vector containing marginal data}

\item{marginal1}{an integer vector containing marginal data}

\item{maxAttempts}{(optional, default=4) number of retries to make if fitting is unsuccessful}
}
\description{
Generate a 2D population matrix
}

