% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pop3}
\alias{pop3}
\title{Generate a 3D population as a list}
\usage{
pop3(marginal0, marginal1, marginal2, maxAttempts = 4L)
}
\arguments{
\item{marginal0}{an integer vector containing marginal data}

\item{marginal1}{an integer vector containing marginal data}

\item{marginal2}{an integer vector containing marginal data}

\item{maxAttempts}{(optional, default=4) number of retries to make if fitting is unsuccessful}
}
\description{
Generate a 3D population as a list
}

