% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob2IntFreq}
\alias{prob2IntFreq}
\title{Generate integer frequencies from discrete probabilities and an overall population.}
\usage{
prob2IntFreq(pIn, pop)
}
\arguments{
\item{pIn}{a numeric vector of state occupation probabilities. Must sum to unity (to within double precision epsilon)}

\item{pop}{the total population}
}
\value{
an integer vector of frequencies that sum to pop, and the RMS difference from the original values.
}
\description{
This function will generate the closest integer vector to the probabilities scaled to the population.
}
\examples{
prob2IntFreq(c(0.1,0.2,0.3,0.4), 11)
}
