% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qis}
\alias{qis}
\title{Multidimensional QIS}
\usage{
qis(indices, marginals, skips = 0L)
}
\arguments{
\item{indices}{an array listing the dimension indices of each marginal as they apply to the seed values}

\item{marginals}{a List of arrays containing marginal data. The sum of elements in each array must be identical}
}
\value{
an object containing: ...
}
\description{
C++ multidimensional Quasirandom Integer Sampling implementation
}
