% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sobolSequence}
\alias{sobolSequence}
\title{Generate Sobol' quasirandom sequence}
\usage{
sobolSequence(dim, n, skip = 0L)
}
\arguments{
\item{dim}{dimensions}

\item{n}{number of variates to sample}

\item{skip}{number of variates to skip (actual number skipped will be largest power of 2 less than k)}
}
\value{
a n-by-d matrix of uniform probabilities in (0,1).
}
\description{
Generate Sobol' quasirandom sequence
}
\examples{
sobolSequence(2, 1000, 1000) # will skip 512 numbers!
}

