% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synthPop}
\alias{synthPop}
\title{[Legacy] Generate a population in n dimensions given n 1-d marginals.}
\usage{
synthPop(marginals)
}
\arguments{
\item{marginals}{a List of n integer vectors containing marginal data (2 <= n <= 12). The sum of elements in each vector must be identical}
}
\value{
an object containing: the population matrix, the occupancy probability matrix, a convergence flag, the chi-squared statistic, p-value, and error value (nonzero if not converged)
}
\description{
Using Quasirandom Integer Without-replacement Sampling (QIWS), this function
generates an n-dimensional population table where elements sum to the input marginals, and supplemental data.
The scope of this function is rather limited (qis is more general), but it is fast.
}
\examples{
synthPop(list(c(1,2,3,4), c(3,4,3)))
}
