% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synthPopR}
\alias{synthPopR}
\title{Generate a correlated population in 2 dimensions given 2 marginals and a flat correlation.}
\usage{
synthPopR(marginals, rho)
}
\arguments{
\item{marginals}{a List of 2 integer vectors containing marginal data. The sum of elements in each vector must be identical}

\item{rho}{correlation}
}
\value{
an object containing: the population matrix, the occupancy probability matrix, a convergence flag, the chi-squared statistic, p-value, and error value (nonzero if not converged)
}
\description{
Using Quasirandom Integer Without-replacement Sampling (QIWS), this function
generates a 2-dimensional population table where elements sum to the input marginals.
}
