% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unitTest}
\alias{unitTest}
\title{Entry point to enable running unit tests within R (e.g. in testthat)}
\usage{
unitTest()
}
\value{
a List containing, number of tests run, number of failures, and any error messages.
}
\description{
Note: Unit tests are only enabled when the package is compiled with the UNIT_TEST flag (see Makevars)
}
\examples{
unitTest()
}
