% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wip_ipf}
\alias{wip_ipf}
\title{Multidimensional IPF}
\usage{
wip_ipf(seed, indices, marginals)
}
\arguments{
\item{seed}{an n-dimensional array of seed values}

\item{indices}{an array listing the dimension indices of each marginal as they apply to the seed values}

\item{marginals}{a List of arrays containing marginal data. The sum of elements in each array must be identical}
}
\value{
an object containing: ...
}
\description{
C++ multidimensional IPF implementation
}
