% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ipf}
\alias{ipf}
\title{IPF}
\usage{
ipf(seedIn, marginals)
}
\arguments{
\item{marginals}{a List of 2 integer vectors containing marginal data. The sum of elements in each vector must be identical}

\item{seed}{seed}
}
\value{
an object containing: the population matrix, the occupancy probability matrix, a convergence flag, the chi-squared statistic, p-value, and error value (nonzero if not converged)
}
\description{
C++ IPF implementation
}
