% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ipf}
\alias{ipf}
\title{IPF}
\usage{
ipf(seed, marginals)
}
\arguments{
\item{seed}{an n-dimensional array of seed values}

\item{marginals}{a List of n integer vectors containing marginal data. The sum of elements in each vector must be identical}
}
\value{
an object containing: ...
}
\description{
C++ IPF implementation
}
