% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qis}
\alias{qis}
\title{Multidimensional QIS}
\usage{
qis(indices, marginals, skips = 0L)
}
\arguments{
\item{indices}{a List of 1-d arrays specifying the dimension indices of each marginal}

\item{marginals}{a List of arrays containing marginal data. The sum of elements in each array must be identical}

\item{skips}{(optional, default 0) number of Sobol points to skip before sampling}
}
\value{
an object containing: the population matrix, the occupancy probability matrix, a convergence flag, chi-square and p-value
TODO examples
}
\description{
C++ multidimensional Quasirandom Integer Sampling implementation
}
