% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qisi}
\alias{qisi}
\title{QIS-IPF}
\usage{
qisi(seed, indices, marginals, skips = 0L)
}
\arguments{
\item{seed}{an n-dimensional array of seed values}

\item{indices}{a List of 1-d arrays specifying the dimension indices of each marginal}

\item{marginals}{a List of arrays containing marginal data. The sum of elements in each array must be identical}

\item{skips}{(optional, default 0) number of Sobol points to skip before sampling}
}
\value{
an object containing: the population matrix, the occupancy probability matrix, a convergence flag, chi-square and p-value
}
\description{
C++ QIS-IPF implementation
}
\examples{
gender=c(51,49)
age=c(17,27,35,21)
seed=array(c(8,7,9,6,10,8,7,9),dim=c(2,4))
synthpop=qisi(seed, list(1,2),list(gender,age))
}
