% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synthPop}
\alias{synthPop}
\title{Generate a population in n dimensions given n marginals}
\usage{
synthPop(marginals, maxAttempts = 4L)
}
\arguments{
\item{marginals}{a List of n integer vectors containing marginal data (2 <= n <= 12)}

\item{maxAttempts}{(optional, default=4) number of retries to make if fitting is unsuccessful}
}
\description{
Generate a population in n dimensions given n marginals
}

