% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unitTest}
\alias{unitTest}
\title{Entry point to enable runing unit tests in R (e.g. in testthat)}
\usage{
unitTest()
}
\value{
number of test failures.
}
\description{
Entry point to enable runing unit tests in R (e.g. in testthat)
}
\examples{
unitTest()
}
